Allow `StrandTrait::unlink()` to be called when no strands have been linked.

+++ src/Kernel/StrandTrait.php
@@ -579,7 +579,9 @@ trait StrandTrait
      */
     public function unlink(Strand $strand)
     {
-        $this->linkedStrands->detach($strand);
+        if ($this->linkedStrands !== null) {
+            $this->linkedStrands->detach($strand);
+        }
     }
 
     /**
